Fixed remider services
git-svn-id: https://svn.iportalen.dk/svn-private/F2007/trunk@318 d9c49cb3-4128-44d0-8390-2fdd9ac6ca64

+++ core/src/test/java/dk/bregnvig/formula1/event/EventServiceTest.java
@@ -31,7 +31,7 @@ public class EventServiceTest extends AbstractDaoTest {
 		
 		private boolean monzaInvoked;
 
-		public int getDelay(Race race) {
+		public long getDelay(Race race) {
 			if (race.getName().equals(monza.getName())) return 300;
 			return -1;
 		}
